first test of a new rectification method

+++ src/CudaImageProcessing.h
@@ -15,6 +15,6 @@ extern "C" void cudaRGBAtoCuda( float *outDevicePtr, uchar4 *inDevicePtr, unsign
 extern "C" void cudaCudatoRGBA( uchar4 *outDevicePtr, float *inDevicePtr, unsigned int imageW, unsigned int imageH, unsigned int pitch );
 extern "C" void cudaTranspose( float *d_dst, size_t dst_pitch, float *d_src, size_t src_pitch, unsigned int width, unsigned int height );
 extern "C" void cudaIntegrate( float*out, float*in, unsigned int width, unsigned int height, unsigned int pitch );
-extern "C" void cudaWarpImage( uchar4 *d_dst, uchar4 *d_src, int imageW, int imageH, float matrix[9] );
+extern "C" void cudaWarpImage( uchar4 *d_dst, uchar4 *d_src, int imageW, int imageH, double matrix[9] );
 
 #endif//__CUDAIMAGEPROCESSING_H__
